+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Quick Look: PDF preview.
+
 ## [0.16.0] - 2026-06-13
 
 ### Added